+++ manager/ExperienceManager.cs
@@ -2,10 +2,10 @@ using Godot;
 using System;
 
 public partial class ExperienceManager : Node {
-	public float CurrentEx { get; private set; } = 0f;
+	public int CurrentEx { get; private set; } = 0;
 	public int CurrentLevel { get; private set; } = 1;
-	public float TargetEx { get; private set; } = 1f;
-	public float TargetExGrowth { get; private set; } = 5f;
+	public int TargetEx { get; private set; } = 1;
+	public int TargetExGrowth { get; private set; } = 5;
 
     [Signal] public delegate void ExperienceUpdatedEventHandler(float current, float target);
     [Signal] public delegate void LevelUpEventHandler(float newLevel);
@@ -21,17 +21,17 @@ public partial class ExperienceManager : Node {
 	}
 	
 
-	private void OnCollectedExperience(float exNum) {
+	private void OnCollectedExperience(int exNum) {
 		IncreasePlayerExperience(exNum);
 	}
 
-	private void IncreasePlayerExperience(float exNum) {
+	private void IncreasePlayerExperience(int exNum) {
         CurrentEx = Mathf.Min(CurrentEx + exNum, TargetEx);
         EmitSignal(SignalName.ExperienceUpdated, CurrentEx, TargetEx);
         if (Mathf.IsZeroApprox(TargetEx - CurrentEx)) {
             CurrentLevel += 1;
             TargetEx += TargetExGrowth;
-            CurrentEx = 0f;
+            CurrentEx = 0;
             EmitSignal(SignalName.ExperienceUpdated, CurrentEx, TargetEx);
             EmitSignal(SignalName.LevelUp, CurrentLevel);
